feat: Update start script to use PORT env var
The `serve` command in the `start` script was previously hardcoded to use `${PORT:-3000}` which would default to 3000 if PORT was not set. This change removes the default value, ensuring that the script strictly uses the value of the PORT environment variable. This allows for more flexible port configuration without unexpected defaults.